store calls update database simultaneously
